requirements.txt
annotated-types==0.7.0
backtesting==0.6.4
beautifulsoup4==4.13.4
bokeh==3.7.3
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
et_xmlfile==2.0.0
fake-useragent==2.2.0
fonttools==4.59.0
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
narwhals==1.48.1
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pillow==11.3.0
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
pyTelegramBotAPI==4.27.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
soupsieve==2.7
squarify==0.4.4
ta-lib==0.6.4
tenacity==9.1.2
tornado==6.5.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vnai==2.1.9
vnstock==3.2.6
vnstock_ezchart==0.0.2
wordcloud==1.9.4
xyzservices==2025.4.0
zipp==3.23.0
